+++ src/elements/Input/tests/index.js
@@ -1,0 +1,5 @@
+import { generateSnapshots } from 'helpers';
+
+import Input from '../index';
+
+generateSnapshots(Input);