+++ src/elements/TextField/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import TextInput from '../index';
-
-generateSnapshots(TextInput);