+++ src/elements/index.js
@@ -1,7 +1,7 @@
 // this is the primary export for Elements
 // the main export for this lib is `src/index.js`
 export { default as Button } from './Button';
-export { default as TextField } from './TextField';
+export { default as Input } from './Input';
 export { default as TextArea } from './TextArea';
 export { default as Switch } from './Switch';
 export { default as Notification } from './Notification';