+++ src/index.d.ts
@@ -8,6 +8,7 @@ import {
   SyntheticEvent,
   ButtonHTMLAttributes,
   HTMLAttributes,
+  AllHTMLAttributes,
 } from 'react';
 import {
   StyledComponent,
@@ -498,14 +499,9 @@ export declare const Switch: FunctionComponent<SwitchProps>;
 
 export declare const TextArea: StyledComponent<'input', Theme>;
 
-export interface TextFieldProps {
-  type: string;
-  placeholder: string;
-  id: string;
-  onChange: (value: number | string) => any;
-}
+export interface InputProps extends AllHTMLAttributes<HTMLInputElement> {}
 
-export declare const TextField: FunctionComponent<TextFieldProps>;
+export declare const Input: FunctionComponent<InputProps>;
 
 type TypographyVariant = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';