+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 
 # [0.7.0] - 28-02-2019
 
- ### Changes
-- Additional user menu added in the header
+### Changes
+- Rename TextField component to Input
+- Add full width/height to CardContent
+- Additional user menu added in Header
+
+### Fixes
+- Update TextField/Input TypeScript definitions
 
 # [0.6.1] - 27-02-2019